requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
colorama==0.4.6
fastapi==0.125.0
greenlet==3.3.0
h11==0.16.0
httptools==0.7.1
idna==3.11
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
SQLAlchemy==2.0.45
sqlmodel==0.0.27
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
passlib[bcrypt]
python-jose[cryptography]
python-multipart
